backend/requirements.txt
Django==3.2
djangorestframework==3.12.4
djoser==2.1.0
Pillow==9.3.0
gunicorn==20.1.0
psycopg2-binary==2.9.3
django-filter